first

+++ core/src/com/sea/battle/lite/playing_field/Cell.java
@@ -3,7 +3,7 @@ package com.sea.battle.lite.playing_field;
 public class Cell {
     byte tip;
 
-    public int getTip() {
+    public byte getTip() {
         return tip;
     }
